Rename modules
- Rename `gpg` to `validate-gpg`
- Rename `orchestrator` to `handle-event`
