+++ lib/handle-event.js
@@ -1,4 +1,4 @@
-const gpg = require('./gpg');
+const validateGpg = require('./validate-gpg');
 const createStatus = require('./create-status');
 
 module.exports = async (robot, event, context) => {
@@ -8,7 +8,7 @@ module.exports = async (robot, event, context) => {
 
   let allCommitsVerified;
   try {
-    allCommitsVerified = await gpg(github, pullRequest.base.sha, headSha);
+    allCommitsVerified = await validateGpg(github, pullRequest.base.sha, headSha);
   } catch (err) {
     allCommitsVerified = 'error';
   }
