+++ lib/validate-gpg.js
