+++ test/handle-event.js
@@ -8,20 +8,20 @@ const RobotMock = require('./mocks/robot');
 const createSha = require('./utils/create-sha');
 const createEvent = require('./utils/create-event');
 
-function arrangeOrchestrator(gpg, createStatus) {
-  return proxyquire('../lib/orchestrator', {
-    './gpg': gpg,
+function arrangeHandler(validateGpg, createStatus) {
+  return proxyquire('../lib/handle-event', {
+    './validate-gpg': validateGpg,
     './create-status': createStatus
   });
 }
 
-describe('orchestrator', () => {
-  async function testScenario(gpgSpy, allCommitsVerified, statusState) {
+describe('handle-event', () => {
+  async function testScenario(validateGpgSpy, allCommitsVerified, statusState) {
     // Arrange
     const createStatusResult = { state: statusState };
     const createStatusSpy = expect.createSpy().andReturn(Promise.resolve(createStatusResult));
 
-    const orchestratorUnderTest = arrangeOrchestrator(gpgSpy, createStatusSpy);
+    const handlerUnderTest = arrangeHandler(validateGpgSpy, createStatusSpy);
 
     const githubMock = new GitHubMock();
     const robotMock = new RobotMock(githubMock);
@@ -33,28 +33,28 @@ describe('orchestrator', () => {
     const contextMock = new ContextMock();
 
     // Act
-    const result = await orchestratorUnderTest(robotMock, event, contextMock);
+    const result = await handlerUnderTest(robotMock, event, contextMock);
 
     // Assert
     expect(robotMock.auth).toHaveBeenCalledWith(event.payload.installation.id);
-    expect(gpgSpy).toHaveBeenCalledWith(githubMock, baseSha, headSha);
+    expect(validateGpgSpy).toHaveBeenCalledWith(githubMock, baseSha, headSha);
     expect(createStatusSpy).toHaveBeenCalledWith(githubMock, contextMock, headSha, allCommitsVerified);
     expect(result).toBe(createStatusResult);
   }
 
   it('should orchestrate correctly when all commits are verified', async () => {
-    const gpgSpy = expect.createSpy().andReturn(true);
-    await testScenario(gpgSpy, true, 'success');
+    const validateGpgSpy = expect.createSpy().andReturn(true);
+    await testScenario(validateGpgSpy, true, 'success');
   });
 
   it('should orchestrate correctly when all commits are not verified', async () => {
-    const gpgSpy = expect.createSpy().andReturn(false);
-    await testScenario(gpgSpy, false, 'failure');
+    const validateGpgSpy = expect.createSpy().andReturn(false);
+    await testScenario(validateGpgSpy, false, 'failure');
   });
 
-  it('should orchestrate correctly when gpg verification check fails', async () => {
+  it('should orchestrate correctly when GPG verification check fails', async () => {
     const err = new Error('The verification status of the commit cannot be determined');
-    const gpgSpy = expect.createSpy().andThrow(err);
-    await testScenario(gpgSpy, 'error', 'error');
+    const validateGpgSpy = expect.createSpy().andThrow(err);
+    await testScenario(validateGpgSpy, 'error', 'error');
   });
 });
